+++ frontend/src/pages/Nodes/actions.js
@@ -7,6 +7,9 @@ export const APPROVE_ORGANIZATION_SUCCESS = "APPROVE_ORGANIZATION_SUCCESS";
 export const APPROVE_NEW_NODE_FOR_ORGANIZATION = "APPROVE_NEW_NODE_FOR_ORGANIZATION";
 export const APPROVE_NEW_NODE_FOR_ORGANIZATION_SUCCESS = "APPROVE_NEW_NODE_FOR_ORGANIZATION_SUCCESS";
 
+export const DECLINE_NODE = "DECLINE_NODE";
+export const DECLINE_NODE_SUCCESS = "DECLINE_NODE_SUCCESS";
+
 export function fetchNodes(showLoading = false) {
   return {
     type: FETCH_NODES,
@@ -25,6 +28,13 @@ export function approveNewNodeForExistingOrganization(address, showLoading = fal
   return {
     type: APPROVE_NEW_NODE_FOR_ORGANIZATION,
     address,
+    showLoading
+  };
+}
+export function declineNode(node, showLoading = false) {
+  return {
+    type: DECLINE_NODE,
+    node,
     showLoading
   };
 }